requirements.txt
asgiref==3.6.0
blinker==1.5
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.1.1
cryptography==37.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django==4.1.5
django-cors-headers==3.13.0
django-crontab==0.7.1
django-phone-field==1.8.1
django-rest-auth==0.9.5
django-upgrade==1.12.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
idna==3.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
Pillow==9.4.0
pip==22.3.1
protobuf==3.20.1
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.3
tokenize-rt==5.0.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
wheel==0.38.4
win-inet-pton==1.1.0
